+++ packages/lua_tool_runtime/lib/lua_tool_runtime.dart
@@ -3,5 +3,6 @@ library;
 
 export 'src/model.dart';
 export 'src/process.dart';
+export 'src/program_bundle.dart';
 export 'src/runtime.dart';
 export 'src/staging.dart';